+++ trunk/azureus2/src/com/aelitis/azureus/core/dht/transport/udp/impl/Test.java
@@ -153,7 +153,7 @@ Test
 								return 0;
 							}
 
-							public int 
+							public byte 
 							getReplicationFactor() 
 							{
 								return 0;
